refactor: renames project page to CurrentProject and implements name change
Relates #36
