+++ src/components/page-components/CurrentProjectPage/CurrentProjectPage.css
@@ -1,21 +1,21 @@
-.project-details {
+.current-project {
   background-color: white;
   animation: slideIn 0.5s ease-out;
 }
 
-.project-details-enter {
+.current-project-enter {
   z-index: 3;
 }
 
-.project-details-enter-active {
+.current-project-enter-active {
   z-index: 3;
 }
 
-.project-details-exit {
+.current-project-exit {
   z-index: 1;
 }
 
-.project-details-exit-active {
+.current-project-exit-active {
   z-index: 1;
 }
