+++ src/popup.tsx
@@ -6,9 +6,9 @@ import { CSSTransition } from "react-transition-group"
 import ExplorePage from "~components/ExplorePage/ExplorePage"
 import HomePage from "~components/HomePage/HomePage"
 import Layout from "~components/Layout/Layout"
-import ProjectDetails from "~components/page-components/ProjectDetails/ProjectDetails"
 import ProfilePage from "~components/ProfilePage/ProfilePage"
 import useStore from "~store/store"
+import CurrentProjectPage from "~components/page-components/CurrentProjectPage/CurrentProjectPage"
 
 function IndexPopup() {
   const currentPage = useStore.use.currentPage()
@@ -43,12 +43,12 @@ function IndexPopup() {
       </CSSTransition>
 
       <CSSTransition
-        in={currentPage === "project-details"}
+        in={currentPage === "current-project"}
         timeout={500}
-        classNames="project-details"
+        classNames="current-project"
         unmountOnExit
       >
-        <ProjectDetails />
+        <CurrentProjectPage />
       </CSSTransition>
     </Layout>
   )
